Changes several Icons to WEBP
This is an extremely annoying and long winded issue. To be as concise as possible, there were rendering issues for many of the icons as svg. Some aren't noticeable but for the user icons, it was awful. Because of this. I decided to switch these icons to webp. This sadly, even after compressing it, still is a much larger file size (ex 4kb to 50kb), but there is no other solution. I tried for hours to get the svgs to render and it seems to be an inherent issue in many browsers.
Thankfully webp is totally fine. It just means some images won't have as insanely low file sizes, but they are still rather small.