Added journalViewJournal page and javascript report download
- Added the view journal html, css and javascript
- This page shows a nicely formatted display showing the user a specific journal and all info related to that journal
- This design is very similar to other reports and will be used for the emotion and strategy reports
- Also added javascript that lets the user download the journal as a text file

NOTE ABOUT DOWNLOAD
I tried for the life of me to make it so the user could just download the whole report as a png since the page nicely formats everything, however, not matter what I tried, it didn't work. I am admittedly not great with javascript and this task sadly was above me. So for now, I have all the info saved to a text file for the user. While this does accomplish the functionality, It's not an ideal solution imo.